refactor: move dns_aid into serve/instrumentator per review feedback
Move dns_aid.py from vllm/entrypoints/ into the instrumentator package
where other server observability modules (health, metrics, offline_docs)
live. Update all import paths in server_utils.py and tests.

Addresses review from @noooop on PR #38513.
